Finished Edit Hotel and Created Edit Reservation.
I Modified only the Hotel Database Manager file to write the function to edit reservations and finished the function to edit the hotel. The edit reservation, however, still needs a way to get the dates. I guess we will discuss that during our next meeting. Anyway, here is the modified file with the edit hotel in the database function ready and the edit reservation function started to show some progress.