Modified Results Class
The hotel results page now contains the ability to store a single hotel
in an object. This allows the program to retrieve other hotel data
besides the name more easily. The user must be able to see the other
data when viewing the hotel, so the initialization of this hotel object
allows the program to call the getters to get a single attribute. Also,
a method has been created to close each of the input streams for the
images of each hotel to prevent resource leaks since Java does not
release those in their garbage collection.