Modified editReservation
I modified the HotelDBManager method: editReservation() so that it can
calculate the new total cost of the reservation. I plan on adding one
more parameter for the previous room type so that that number of rooms
can be updated as well. There are still many things I need to do to
protect the integrity of the number of rooms available, which I will try
to tackle by the end of this week.