Added FIXME statement in search() method.
I only added this FIXME statement so that I can refer back to it. The
purpose is to modify the results query so that hotels without a king or
queen bed (rmPrice for each = $0.00) can use their next highest price
when given a maximum price.

I also plan to add some other changes like returning a code in the
bookReservation method where there is not enough rooms of a room type
available. this is to prevent a negative number of rooms from occurring.