Updated Book Reservation.
The book reservation method now checks the number of rooms available to
check that the user's entered number of rooms does not exceed that
amount before updating any of the tables or adding their reservation to
the Reservation table.

I am also still trying to figure out the bug while going back to a hotel
and load its image from the search results.