Fixed prices and started working on checking database integrity.
I did the price calculations of the reservations for my user by hand and
the costs when I see the reservation come out correct, so that is
finally over with! Also because of this, I removed some classes and
methods I did not need to use anymore since I was fiddling around with
the dates a lot, and I finally came up with a simple solution, well,
somewhat simple.